Completed technical part
Well, now I have finished the entire technical part of the admin bot. At the moment, all the mechanisms that are in this bot are debugged and working properly. All the bugs that I could imagine in my head have been fixed. Yes, my bot is not perfect yet, there are a lot of so-called spelling errors). But the most important thing is that it works. And I'm incredibly happy about it. I created this bot for almost exactly three months. I reworked it a bunch of times and finally I can say that the most difficult part is finished. And minor errors and improvements to the code will occur over time.